backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
gunicorn==21.2.0
pydantic==2.8.2
pydantic-settings==2.2.1
python-dotenv==1.0.0

# Bot Framework SDK for Python
botbuilder-core==4.15.0
botbuilder-schema==4.15.0
botbuilder-integration-aiohttp==4.15.0

# LangChain/LangGraph (0.3 line)
langchain>=0.3,<0.4
langchain-core>=0.3.43,<0.4
langgraph>=0.2.27,<0.3

# Gemini integration on 0.3 core + new SDK
langchain-google-genai>=2.1.0,<2.2
google-genai>=1.18.0,<2.0

# Azure
azure-search-documents==11.6.0b4
azure-identity==1.17.1
azure-core==1.30.2
azure-storage-blob==12.16.0

# Data & utils
pandas==2.2.2
openpyxl==3.1.5
httpx==0.28.1
